Port WNDCLASS (#7573)
Ports `WNDCLASS` and style definitions. Convert `NMHDR`.

Port `LoadCursor`, `LoadIcon`, `ShowCursor`, and `DestroyWindow`.

Ports `CreateWindowEx` and manually implments `AsAny` marshalling. Note that arbitrary arrays are not supported in this change, but simple arrays of intrinsic types still are (`int[]`).

Adds `IHandle<HWND>` to `HWND` to reduce the need for multiple overloads (see `IsChild` and `SetParent`).

# Conflicts:
#	src/System.Windows.Forms.Primitives/src/NativeMethods.txt

+++ src/System.Windows.Forms/src/System/Windows/Forms/FileDialog.cs
@@ -534,7 +534,7 @@ namespace System.Windows.Forms
                 {
                     Comdlg32.OFNOTIFYW* notify = (Comdlg32.OFNOTIFYW*)lparam;
 
-                    switch (notify->hdr.code)
+                    switch ((int)notify->hdr.code)
                     {
                         case -601: /* CDN_INITDONE */
                             MoveToScreenCenter(_dialogHWnd);
